feat: Adde throbber and improved default search view styles
