+++ sanitas_frontend/src/components/Throbber/index.stories.jsx
@@ -1,0 +1,17 @@
+import React from "react";
+import Throbber from ".";
+
+export default {
+  title: "Components/Throbber",
+  component: Throbber,
+};
+
+const Template = (args) => <Throbber {...args} />;
+
+export const Default = Template.bind({});
+/**
+ * @type {import(".").ThrobberProps}
+ */
+Default.args = {
+  loadingMessage: "Cargando data...",
+};
